modification de la fonction eventKeys

+++ src/common/keys.h
@@ -6,12 +6,12 @@ typedef struct{
     int SDLK;
 }Keys;
 
-int charToSdlk(char * key);
+uint16_t charToSdlk(char * key);
 /* Translated string in SDLK
    Input : char * key :  the name of the key
    Output : Value of the SDLK_*/
 
-char * SdlkToChar(int sdlk);
+char * SdlkToChar(uint16_t sdlk);
 /* Translated SDLK in string
    Input : int sdlk :  SDLK code for a key
    Output :  string */
@@ -29,7 +29,7 @@ int readFileKeys(char * nameFile, uint16_t * keysSelect);
    Output : 1 or 0*/
 
 
-int eventKeys(uint16_t * keysSelect);
+uint8_t eventKeys(uint16_t * keysSelect);
 /* detect a event
     Input :  int * keysSelect : Table containing the SDLK keys
     Ouput : a value in 1 byte*/
